vulnmonolith/services/billing-service/requirements.txt
# Billing Service Requirements
# WARNING: Contains intentionally vulnerable package versions

# Web Framework
fastapi==0.68.0
uvicorn==0.15.0
starlette==0.14.2

# Database
sqlalchemy==1.3.0
databases==0.4.0
asyncpg==0.22.0
psycopg2-binary==2.8.6

# VULNERABLE: PyYAML with arbitrary code execution
PyYAML==5.3

# VULNERABLE: Pillow with multiple CVEs
Pillow==8.0.0

# VULNERABLE: requests with CVE
requests==2.20.0

# VULNERABLE: urllib3
urllib3==1.24.1

# VULNERABLE: Jinja2 SSTI
Jinja2==2.10

# Serialization (for pickle vulnerabilities)
dill==0.3.3

# Authentication
python-jose==3.2.0
passlib==1.7.4
bcrypt==3.2.0

# Validation
pydantic==1.8.0
email-validator==1.1.3

# Payment Processing
stripe==2.55.0

# Utilities
python-multipart==0.0.5
python-dotenv==0.19.0
aiofiles==0.7.0
httpx==0.18.0

# VULNERABLE: cryptography (older version)
cryptography==3.3.1

# Templating
mako==1.1.0

# XML Processing (XXE potential)
lxml==4.6.2
defusedxml==0.7.1

# Data Processing
pandas==1.2.0
numpy==1.19.0
